# Enable int8xint8->int32 `einsum`