--- a/keras/backend/tensorflow/numpy.py
+++ b/keras/backend/tensorflow/numpy.py
@@ -1,6 +1,8 @@
 import builtins
 import collections
+import functools
 import math
+import string
 import warnings
 
 import numpy as np
@@ -75,22 +77,105 @@ def bincount(x, weights=None, minlength=0):
     )
 
 
+@functools.lru_cache(512)
+def _normalize_einsum_subscripts(subscripts):
+    # string.ascii_letters
+    mapping = {}
+    normalized_subscripts = ""
+    for c in subscripts:
+        if c in string.ascii_letters:
+            if c not in mapping:
+                mapping[c] = string.ascii_letters[len(mapping)]
+            normalized_subscripts += mapping[c]
+        else:
+            normalized_subscripts += c
+    return normalized_subscripts
+
+
 def einsum(subscripts, *operands, **kwargs):
     operands = tf.nest.map_structure(convert_to_tensor, operands)
+    subscripts = _normalize_einsum_subscripts(subscripts)
 
-    dtypes_to_resolve = []
-    for x in operands:
-        dtypes_to_resolve.append(x.dtype)
-    result_dtype = dtypes.result_type(*dtypes_to_resolve)
-    compute_dtype = result_dtype
-    # TODO: tf.einsum doesn't support integer dtype with gpu
-    if "int" in compute_dtype:
-        compute_dtype = config.floatx()
+    def is_valid_for_custom_ops(subscripts, *operands):
+        # Check that `subscripts` is supported and the shape of operands is not
+        # `None`
+        if subscripts == "abcde,afce->acdbf":
+            _, b1, c1, d1, e1 = operands[0].shape
+            _, f2, c2, e2 = operands[1].shape
+            b, c, d, e, f = b1, c1 or c2, d1, e1 or e2, f2
+            if None in (b, c, d, e, f):
+                return False
+            return True
+        elif subscripts == "abc,dce->abde":
+            _, b1, c1 = operands[0].shape
+            d2, c2, e2 = operands[1].shape
+            b, c, d, e = b1, c1 or c2, d2, e2
+            if None in (b, c, d, e):
+                return False
+            return True
+        else:
+            # Defaults to `False`
+            return False
 
-    operands = tf.nest.map_structure(
-        lambda x: tf.cast(x, compute_dtype), operands
-    )
-    return tf.cast(tf.einsum(subscripts, *operands, **kwargs), result_dtype)
+    def use_custom_ops(subscripts, *operands, output_type):
+        # Replace tf.einsum with custom ops to utilize hardware-accelerated
+        # matmul
+        if subscripts == "abcde,afce->acdbf":
+            x = operands[0]
+            y = operands[1]
+            _, b1, c1, d1, e1 = x.shape
+            _, f2, c2, e2 = y.shape
+            b, c, d, e, f = b1, c1 or c2, d1, e1 or e2, f2
+            x = tf.transpose(x, [0, 2, 3, 1, 4])  # acdbe
+            x = tf.reshape(x, [-1, c, d * b, e])  # ac(b*d)e
+            y = tf.transpose(y, [0, 2, 3, 1])  # acef
+            result = tf.matmul(x, y, output_type=output_type)
+            return tf.reshape(result, [-1, c, d, b, f])
+        elif subscripts == "abc,dce->abde":
+            x = operands[0]
+            y = operands[1]
+            _, b1, c1 = x.shape
+            d2, c2, e2 = y.shape
+            b, c, d, e = b1, c1 or c2, d2, e2
+            x = tf.reshape(x, [-1, c])
+            y = tf.transpose(y, [1, 0, 2])
+            y = tf.reshape(y, [c, -1])
+            result = tf.matmul(x, y, output_type=output_type)
+            return tf.reshape(result, [-1, b, d, e])
+        else:
+            raise NotImplementedError
+
+    dtypes_to_resolve = list(set(standardize_dtype(x.dtype) for x in operands))
+    # When operands are of int8, we cast the result to int32 to align with
+    # the behavior of jax.
+    if len(dtypes_to_resolve) == 1 and dtypes_to_resolve[0] == "int8":
+        compute_dtype = "int8"
+        result_dtype = "int32"
+        output_type = "int32"
+    else:
+        result_dtype = dtypes.result_type(*dtypes_to_resolve)
+        compute_dtype = result_dtype
+        output_type = None
+
+    # TODO: Remove the condition once `tf.einsum` supports int8xint8->int32
+    if is_valid_for_custom_ops(subscripts, *operands) and not kwargs:
+        # TODO: tf.matmul doesn't support integer dtype if not specifying
+        # output_type="int32"
+        if "int" in compute_dtype and output_type is None:
+            compute_dtype = config.floatx()
+        operands = tf.nest.map_structure(
+            lambda x: tf.cast(x, compute_dtype), operands
+        )
+        result = use_custom_ops(subscripts, *operands, output_type=output_type)
+    else:
+        # TODO: tf.einsum doesn't support integer dtype with gpu
+        if "int" in compute_dtype:
+            compute_dtype = config.floatx()
+        operands = tf.nest.map_structure(
+            lambda x: tf.cast(x, compute_dtype), operands
+        )
+        result = tf.einsum(subscripts, *operands, **kwargs)
+    return tf.cast(result, result_dtype)
 
 
 @sparse.elementwise_binary_union(sparse.sparse_subtract)