--- a/keras/backend/torch/numpy.py
+++ b/keras/backend/torch/numpy.py
@@ -29,6 +29,13 @@ def add(x1, x2):
 
 def einsum(subscripts, *operands, **kwargs):
     operands = [convert_to_tensor(operand) for operand in operands]
+    # When all operands are of int8, we cast the result to int32 to align with
+    # the behavior of jax.
+    dtypes_to_resolve = list(set(standardize_dtype(x.dtype) for x in operands))
+    if len(dtypes_to_resolve) == 1 and dtypes_to_resolve[0] == "int8":
+        # prevent overflow
+        operands = [cast(operand, "int32") for operand in operands]
+        return cast(torch.einsum(subscripts, *operands), "int32")
     return torch.einsum(subscripts, *operands)
 
 
@@ -43,38 +50,37 @@ def subtract(x1, x2):
     return torch.subtract(x1, x2)
 
 
-def _can_use_int_matmul(x1, x2):
-    # torch._int_mm only accepts the following conditions:
-    # 1. cuda
-    # 2. both inputs must have int8 dtype
-    # 3. both inputs must be 2d
-    # 4. x1.shape must be [>16, >= 16 and a multiplier of 8]
-    # 5. x2.shape must be [>= 16 and a multiplier of 8, multiplier of 8]
-    if get_device() != "cuda":
-        return False
-    x1_dtype = standardize_dtype(x1.dtype)
-    x2_dtype = standardize_dtype(x2.dtype)
-    if x1_dtype != "int8" or x2_dtype != "int8":
-        return False
-    x1_shape = x1.shape
-    x2_shape = x2.shape
-    if x1.ndim != 2 or x2.ndim != 2:
-        return False
-    if x1_shape[0] <= 16 or x1_shape[1] < 16 or x1_shape[1] % 8 != 0:
-        return False
-    if x2_shape[0] < 16 or x2_shape[0] % 8 != 0 or x2_shape[1] % 8 != 0:
-        return False
-    return True
-
-
 def matmul(x1, x2):
     x1 = convert_to_tensor(x1)
     x2 = convert_to_tensor(x2)
 
+    def can_use_int_matmul(x1, x2):
+        # torch._int_mm only accepts the following conditions:
+        # 1. cuda
+        # 2. both inputs must have int8 dtype
+        # 3. both inputs must be 2d
+        # 4. x1.shape must be [>16, >= 16 and a multiplier of 8]
+        # 5. x2.shape must be [>= 16 and a multiplier of 8, multiplier of 8]
+        if get_device() != "cuda":
+            return False
+        x1_dtype = standardize_dtype(x1.dtype)
+        x2_dtype = standardize_dtype(x2.dtype)
+        if x1_dtype != "int8" or x2_dtype != "int8":
+            return False
+        x1_shape = x1.shape
+        x2_shape = x2.shape
+        if x1.ndim != 2 or x2.ndim != 2:
+            return False
+        if x1_shape[0] <= 16 or x1_shape[1] < 16 or x1_shape[1] % 8 != 0:
+            return False
+        if x2_shape[0] < 16 or x2_shape[0] % 8 != 0 or x2_shape[1] % 8 != 0:
+            return False
+        return True
+
     # Shortcut for torch._int_mm
     # TODO: Loosen the restriction of the usage of torch._int_mm
     # TODO: We should replace torch._int_mm with the public api if possible
-    if _can_use_int_matmul(x1, x2):
+    if can_use_int_matmul(x1, x2):
         return torch._int_mm(x1, x2)
 
     x1_dtype = standardize_dtype(x1.dtype)