--- a/keras/ops/numpy_test.py
+++ b/keras/ops/numpy_test.py
@@ -2136,6 +2136,23 @@ class NumpyTwoInputOpsCorretnessTest(testing.TestCase, parameterized.TestCase):
         )
         self.assertAllClose(knp.Einsum(",ijk")(5, y), np.einsum(",ijk", 5, y))
 
+    def test_einsum_with_custom_ops(self):
+        # abcde,afce->acdbf
+        x = np.arange(720).reshape([2, 3, 4, 5, 6]).astype("float32")
+        y = np.arange(336).reshape([2, 7, 4, 6]).astype("float32")
+        self.assertAllClose(
+            knp.einsum("abcde,afce->acdbf", x, y),
+            np.einsum("abcde,afce->acdbf", x, y),
+        )
+
+        # abc,dce->abde
+        x = np.arange(24).reshape([2, 3, 4]).astype("float32")
+        y = np.arange(120).reshape([5, 4, 6]).astype("float32")
+        self.assertAllClose(
+            knp.einsum("abc,dce->abde", x, y),
+            np.einsum("abc,dce->abde", x, y),
+        )
+
     def test_full_like(self):
         x = np.array([[1, 2, 3], [3, 2, 1]])
         self.assertAllClose(knp.full_like(x, 2), np.full_like(x, 2))
@@ -5452,7 +5469,10 @@ class NumpyDtypeTest(testing.TestCase, parameterized.TestCase):
         self.assertEqual(knp.Dot().symbolic_call(x1, x2).dtype, expected_dtype)
 
     @parameterized.named_parameters(
-        named_product(dtypes=itertools.combinations(ALL_DTYPES, 2))
+        named_product(
+            dtypes=list(itertools.combinations(ALL_DTYPES, 2))
+            + [("int8", "int8")]
+        )
     )
     def test_einsum(self, dtypes):
         import jax.numpy as jnp
@@ -5462,9 +5482,18 @@ class NumpyDtypeTest(testing.TestCase, parameterized.TestCase):
         x2 = knp.ones((1, 1, 1), dtype=dtype2)
         x1_jax = jnp.ones((1, 1, 1), dtype=dtype1)
         x2_jax = jnp.ones((1, 1, 1), dtype=dtype2)
+        if dtype1 == "int8" and dtype2 == "int8":
+            preferred_element_type = "int32"
+        else:
+            preferred_element_type = None
         subscripts = "ijk,lkj->il"
         expected_dtype = standardize_dtype(
-            jnp.einsum(subscripts, x1_jax, x2_jax).dtype
+            jnp.einsum(
+                subscripts,
+                x1_jax,
+                x2_jax,
+                preferred_element_type=preferred_element_type,
+            ).dtype
         )
 
         self.assertEqual(
@@ -5477,6 +5506,54 @@ class NumpyDtypeTest(testing.TestCase, parameterized.TestCase):
             ),
             expected_dtype,
         )
+
+        # Test custom implementation of einsum for tensorflow
+        if backend.backend() == "tensorflow":
+            subscripts = "abc,dce->abde"
+            expected_dtype = standardize_dtype(
+                jnp.einsum(
+                    subscripts,
+                    x1_jax,
+                    x2_jax,
+                    preferred_element_type=preferred_element_type,
+                ).dtype
+            )
+
+            self.assertEqual(
+                standardize_dtype(knp.einsum(subscripts, x1, x2).dtype),
+                expected_dtype,
+            )
+            self.assertEqual(
+                standardize_dtype(
+                    knp.Einsum(subscripts).symbolic_call(x1, x2).dtype
+                ),
+                expected_dtype,
+            )
+
+            subscripts = "abcde,afce->acdbf"
+            x1 = knp.ones((1, 1, 1, 1, 1), dtype=dtype1)
+            x2 = knp.ones((1, 1, 1, 1), dtype=dtype2)
+            x1_jax = jnp.ones((1, 1, 1, 1, 1), dtype=dtype1)
+            x2_jax = jnp.ones((1, 1, 1, 1), dtype=dtype2)
+            expected_dtype = standardize_dtype(
+                jnp.einsum(
+                    subscripts,
+                    x1_jax,
+                    x2_jax,
+                    preferred_element_type=preferred_element_type,
+                ).dtype
+            )
+
+            self.assertEqual(
+                standardize_dtype(knp.einsum(subscripts, x1, x2).dtype),
+                expected_dtype,
+            )
+            self.assertEqual(
+                standardize_dtype(
+                    knp.Einsum(subscripts).symbolic_call(x1, x2).dtype
+                ),
+                expected_dtype,
+            )
 
     @parameterized.named_parameters(named_product(dtype=ALL_DTYPES))
     def test_empty(self, dtype):