jax:
use `preferred_element_type="int32"` to enable int8xint8->int32 in `einsum`

tf:
add custom ops for 2 subscripts below to replace `tf.einsum` with `tf.matmul` in order to enable int8xint8->int32 format:
- "abcde,afce->acdbf"
- "abc,dce->abde"

other backends:
just align the dtype conversion with jax and tf.

However, I'm not sure how to benchmark them, and currently, I don't see much improvement by using int8 `einsum` compared to float32/bfloat16 `einsum` in jax and tf backends.