@fchollet 2024-02-28
Thanks for the PR!

> However, I'm not sure how to benchmark them, and currently, I don't see much improvement by using int8 einsum compared to float32/bfloat16 einsum in jax and tf backends.

You can simply create a `@jax.jit` or `@tf.function(jit_compile=True)` function that calls einsum, and call it repeatedly it for large random arrays with the 2 supported subscripts. Large would be something in line with array sizes in LLMs, e.g. `(32, 512, 512)` for the input. Make sure to not count the initial compilation time (so only count after the 1st execution).

What numbers are you currently seeing for different backends?