@james77777778 2024-02-29
@fchollet 
Thank you for the reasonable settings!

I've adjusted the benchmark script from #19137  to ensure that the computation will utilize nearly 100% of my 4070.

```python
import time

from keras import backend
from keras import ops
from keras import random

SUBSCRIPTS = "abc,dce->abde"
# [b, dim, dim], [dim, dim, dim] -> [b, dim, dim, dim]


def int8_linear_layer(x, params):
    # read in x and w as pre-quantized int8 tensors
    y = ops.einsum(SUBSCRIPTS, x, params["w"])

    # add bias and apply activation in fp32
    y = ops.cast(y, "float32")
    y = y + params["b"]
    y = ops.relu(y)

    # quantize and store output as int8
    y = ops.round(y / params["s"])
    y = ops.clip(y, -128, 127)
    y = ops.cast(y, "int8")
    y = y[..., 0]
    return y


def fp_linear_layer(x, params):
    y = ops.relu(ops.add(ops.einsum(SUBSCRIPTS, x, params["w"]), params["b"]))
    y = y[..., 0]
    return y


def linear_layer(x, params):
    if backend.standardize_dtype(params["w"].dtype) == "int8":
        return int8_linear_layer(x, params)
    else:
        return fp_linear_layer(x, params)


def network(x, network_params):
    for layer_params in network_params:
        x = linear_layer(x, layer_params)
    return x[0, 0]


def benchmark(batch_size, dim, dtype, n_layers=32, iters=200):
    """Benchmarks the throughput of a linear layer.

    matmul + bias + act (+ quantization)

    To drown out data transfer costs and other overheads, each run performs a
    forward pass of a network consisting of `n_layers` of linear layers in
    series.

    The full network is run for `iters` iterations to compute the throughput in
    TFLOPS/TOPS.
    """

    # create input tensor for network
    if dtype == "int8":
        x = random.uniform(
            [batch_size, dim, dim], minval=-127, maxval=127, dtype="float32"
        )
        x = ops.cast(x, "int8")
    else:
        x = random.uniform([batch_size, dim, dim], dtype=dtype)

    # create weight + bias (+ quant scale) tensors for each layer
    network_params = []
    for i in range(n_layers):
        if dtype == "int8":
            w = random.uniform(
                [dim, dim, dim], minval=-127, maxval=127, dtype="float32"
            )
            w = ops.cast(w, "int8")
            b = random.normal([dim], dtype="float32")
            s = random.normal([], dtype="float32")
            layer_params = {"w": w, "b": b, "s": s}
        else:
            w = random.uniform([dim, dim, dim], dtype=dtype)
            b = random.uniform([dim], dtype=dtype)
            layer_params = {"w": w, "b": b}
        network_params.append(layer_params)

    # jit
    if backend.backend() == "tensorflow":
        import tensorflow as tf

        jit_network = tf.function(network, jit_compile=True)
    elif backend.backend() == "jax":
        import jax

        jit_network = jax.jit(network)
    elif backend.backend() == "torch":
        import torch

        torch.set_float32_matmul_precision("high")
        jit_network = torch.compile(network)

    # warmup
    for _ in range(10):
        _ = ops.convert_to_numpy(jit_network(x, network_params))

    times = []
    for _ in range(iters):
        t0 = time.time()
        _ = ops.convert_to_numpy(jit_network(x, network_params))
        elapsed_time = time.time() - t0  # in ms
        times.append(elapsed_time)

    avg_time = sum(times) / len(times)
    gflops = (n_layers * 2 * dim**3) / avg_time * 1e-9
    return gflops


if __name__ == "__main__":
    batch_size = 16

    # not used, just triggers tensorflow/xla initial log dump
    benchmark(batch_size, 64, "float16")

    print("\nGOPS/GFLOPS")
    cols = f"{'dim':<6} | {'int8':>8} {'fp16':>8} {'fp32':>8}"
    print("=" * len(cols) + "\n" + cols + "\n" + "-" * len(cols))
    for dim in [32, 64, 128, 256]:
        int8_time = benchmark(batch_size, dim, "int8", 16)
        fp16_time = benchmark(batch_size, dim, "float16", 16)
        fp32_time = benchmark(batch_size, dim, "float32", 16)
        print(
            f"{dim:<6} | {int8_time:>8.2f} {fp16_time:>8.2f} {fp32_time:>8.2f}"
        )
    print("=" * len(cols))

```

And the results:
```bash
# tensorflow
GOPS/GFLOPS
===================================
dim    |     int8     fp16     fp32
-----------------------------------
32     |     1.82     2.10     1.93
64     |     9.98     9.20     6.56
128    |    10.62     9.15     4.14
256    |    11.71     8.04     3.97
===================================

# jax
GOPS/GFLOPS
===================================
dim    |     int8     fp16     fp32
-----------------------------------
32     |     6.47     6.07     4.81
64     |    19.31    14.84     8.56
128    |    11.18     9.46     4.19
256    |    11.85     8.10     3.98
===================================
```

Note that the performance of int8 is not as efficient as the results of `matmul` but it still the best one.