@fchollet 2024-02-29
Nice -- these numbers are in the right ballpark!

> Do we need more subscripts for tensorflow?
> I'm not sure if the 2 included ones are sufficient.

Can you check BERT, GPT-2, Mistral and Gemma from KerasNLP? Just instantiate the model, grab a MHA layer from the decoder blocker (it is usually a custom layer with "attention" in the name), and print `self.equation` for the `query_dense`, `value_dense`, `key_dense` and `output_dense` layers. They may be named e.g. `_query_dense`/etc.

If we cover the equations used by these layers in these models, then we are good!

In many cases it's possible to figure out the equations by looking at the code, without instantiating the model, e.g. https://github.com/keras-team/keras-nlp/blob/master/keras_nlp/models/gemma/gemma_attention.py

- "btd,ndh->btnh"
- "bsd,kdh->bskh"
- "bsd,kdh->bskh"
- "btnh,nhd->btd"