@james77777778 2024-03-01
> LGTM -- the code looks good! If the feature is ready, then we can merge.

I have verified that there is no performance regression after the changes (all done in `"float32"`):

```bash
# master branch
subscripts=a,b->ab           (dim=128), cost time=0.10466ms(±0.00062)
subscripts=abc,cd->abd       (dim=128), cost time=0.11970ms(±0.00055)
subscripts=abc,cde->abde     (dim=128), cost time=0.12020ms(±0.00050)
subscripts=abc,dce->abde     (dim=128), cost time=1.35215ms(±1.30167)
subscripts=abcd,abed->abce   (dim=64 ), cost time=0.14666ms(±0.00067)
subscripts=abcd,adbe->acbe   (dim=64 ), cost time=0.78262ms(±0.37625)
subscripts=abcd,aecd->acbe   (dim=64 ), cost time=0.79861ms(±0.37621)
subscripts=abcd,aecd->aceb   (dim=64 ), cost time=0.79831ms(±0.37650)
subscripts=abcd,cde->abe     (dim=128), cost time=1.33974ms(±1.29205)
subscripts=abcde,aebf->adbcf (dim=32 ), cost time=0.93679ms(±0.44432)
subscripts=abcde,afce->acdbf (dim=32 ), cost time=1.74974ms(±0.17599)
```

```bash
# this pr
subscripts=a,b->ab           (dim=128), cost time=0.10347ms(±0.00054)
subscripts=abc,cd->abd       (dim=128), cost time=0.11793ms(±0.00059)
subscripts=abc,cde->abde     (dim=128), cost time=0.11889ms(±0.00060)
subscripts=abc,dce->abde     (dim=128), cost time=1.36989ms(±1.32058)
subscripts=abcd,abed->abce   (dim=64 ), cost time=0.11818ms(±0.00048)
subscripts=abcd,adbe->acbe   (dim=64 ), cost time=0.79939ms(±0.37633)
subscripts=abcd,aecd->acbe   (dim=64 ), cost time=0.79967ms(±0.37628)
subscripts=abcd,aecd->aceb   (dim=64 ), cost time=0.79946ms(±0.37636)
subscripts=abcd,cde->abe     (dim=128), cost time=1.34044ms(±1.29371)
subscripts=abcde,aebf->adbcf (dim=32 ), cost time=0.93809ms(±0.44360)
subscripts=abcde,afce->acdbf (dim=32 ), cost time=0.95274ms(±0.44201)
```

In `subscripts="abcde,afce->acdbf"`, the custom implementation get even faster performance than pure `tf.einsum`.

The benchmark script:
<details>

```python
import time

import tensorflow as tf

from keras import ops


def get_inputs(subscripts, dim=64):
    x1_labels = subscripts.split(",")[0]
    x2_labels = subscripts.split("->")[0][len(x1_labels) + 1 :]
    x1_shape = [dim] * len(x1_labels)
    x2_shape = [dim] * len(x2_labels)
    x1 = ops.ones(x1_shape)
    x2 = ops.ones(x2_shape)
    return x1, x2


def benchmark(subscripts, iterations=1000):
    if len(subscripts) >= len("abcde,afce->acdbf"):
        dim = 32
    elif len(subscripts) >= len("abcd,abed->abce"):
        dim = 64
    else:
        dim = 128
    x1, x2 = get_inputs(subscripts, dim=dim)

    @tf.function(jit_compile=True)
    def fn(x1, x2):
        return ops.einsum(subscripts, x1, x2)

    # warmup
    for _ in range(20):
        _ = fn(x1, x2)
    cost_times = []
    for _ in range(iterations):
        t0 = time.time()
        _ = fn(x1, x2)
        cost_times.append(time.time() - t0)
    cost_times = ops.convert_to_tensor(cost_times)
    cost_times = ops.sort(cost_times)
    cost_times = cost_times[int(iterations * 0.25) : -int(iterations * 0.25)]
    mean = ops.convert_to_numpy(ops.mean(cost_times)) * 1000.0
    std = ops.convert_to_numpy(ops.std(cost_times)) * 1000.0
    print(
        f"subscripts={subscripts:<17} (dim={dim:<3}), "
        f"cost time={mean:.5f}ms(±{std:.5f})"
    )


for subscripts in [
    "a,b->ab",
    "abc,cd->abd",
    "abc,cde->abde",
    "abc,dce->abde",
    "abcd,abed->abce",
    "abcd,adbe->acbe",
    "abcd,aecd->acbe",
    "abcd,aecd->aceb",
    "abcd,cde->abe",
    "abcde,aebf->adbcf",
    "abcde,afce->acdbf",
]:
    benchmark(subscripts)

```

</details>

I think this PR is ready now.
@fchollet